+++ be-api/src/docms/bootstrap.go
@@ -204,6 +204,7 @@ func _loadDocumentsForTopic(topicMeta *TopicMeta) {
 	if err != nil {
 		panic(err)
 	}
+	topicMeta.numDocs = len(docDirList)
 	for _, docDir := range docDirList {
 		docDirPath := gDataDir + "/" + topicMeta.dir + "/" + docDir.Name()
 		log.Printf("[%s] Loading Document data from <%s>...", logLevelInfo, docDirPath)
@@ -213,10 +214,32 @@ func _loadDocumentsForTopic(topicMeta *TopicMeta) {
 		if err != nil {
 			panic(err)
 		}
-		topicDocId := topicMeta.id + ":" + docMeta.id
+		{
+			// temp fix
+			if docMeta.TimestampUpdate <= 0 {
+				fi, _ := os.Stat(docDirPath)
+				docMeta.TimestampUpdate = fi.ModTime().Unix()
+			}
+			if docMeta.Author == nil {
+				docMeta.Author = gSiteMeta.Author
+			}
+			if docMeta.Author == nil {
+				docMeta.Author = &Author{
+					Name:  goapi.AppConfig.GetString("app.shortname"),
+					Email: goapi.AppConfig.GetString("app.shortname") + "@domain.com",
+				}
+			}
+			if docMeta.Author.Name == "" {
+				docMeta.Author.Name = goapi.AppConfig.GetString("app.shortname")
+			}
+			if docMeta.Author.Email == "" {
+				docMeta.Author.Email = goapi.AppConfig.GetString("app.shortname") + "@domain.com"
+			}
+		}
 		docMeta.setDirectory(docDir.Name())
+		topicDocId := topicMeta.id + ":" + docMeta.id
 		gDocumentListPerTopic[topicMeta.id] = append(gDocumentListPerTopic[topicMeta.id], docMeta)
-		gDocumentList = append(gDocumentList, docMeta)
+		gDocumentList = append(gDocumentList, fmt.Sprintf("%d", docMeta.index)+"/"+topicDocId)
 		gDocumentMeta[topicDocId] = docMeta
 		gDocumentContent[topicDocId] = make(map[string]string)
 		if docMeta.DocPage != "" {
@@ -313,8 +336,10 @@ func _loadTopics() {
 	})
 	if gSiteMeta.Mode == SiteModeBlog {
 		sort.Slice(gDocumentList, func(i, j int) bool {
-			return gDocumentList[i].index > gDocumentList[j].index
+			return gDocumentList[i] > gDocumentList[j]
 		})
+	} else {
+		gDocumentList = make([]string, 0)
 	}
 }
 
@@ -363,6 +388,7 @@ func _reloadCMSData() {
 func initApiHandlers(router *itineris.ApiRouter) {
 	router.SetHandler("getSiteMeta", apiGetSiteMeta)
 	router.SetHandler("getTopics", apiGetTopics)
+	router.SetHandler("getDocuments", apiGetDocuments)
 	router.SetHandler("getDocumentsForTopic", apiGetDocumentsForTopic)
 	router.SetHandler("getDocument", apiGetDocument)
 	router.SetHandler("search", apiSearch)