+++ fe/src/tpl-bootstrap/components/Search.vue
@@ -69,7 +69,7 @@
 </template>
 
 <script>
-import {swichLanguage} from "@/_shared/i18n"
+import {switchLanguage} from "@/_shared/i18n"
 import {watch} from 'vue'
 import {useRoute} from "vue-router"
 import legoPageHeader from './_pageHeader.vue'
@@ -94,7 +94,7 @@ export default {
         () => route.query.l,
         async () => vue._search(vue),
     )
-    swichLanguage(this.searchLocale)
+    switchLanguage(this.searchLocale)
     this.$global.searchQuery = this.searchTerm
     this._fetchSiteMeta(this)
   },